更新v2.3
## [v2.3] - 2020-02-15
* 1.文件管理器添加路径显示
* 2.更新HardwareSerial库
* 3.Page_GetAppWindow更名AppWindow_GetCont
* 4.Module更名BSP
* 5.添加GameDef寄存器级别优化屏幕操作，显著提升游戏模拟器帧率，CPU占用率从70%降至18%
* 6.添加Shell模块，类LINUX操作
* 7.编写SerialUploader上位机，支持透传SD卡文件，CRC32校验
* 8.优化WAV播放器体验
* 9.更新CommonMacro库
* 10.sys_io更名rt_sys，添加注释
* 11.修复栈大小显示错误的BUG
* 12.彻底消灭DisplayPrivate全局变量

+++ Master/XC-OS/Basic/TasksManage.h
@@ -12,7 +12,7 @@
 /*Ext Function*/
 #define DEBUG_PRINTF(format, ...) Serial.printf(format, ##__VA_ARGS__)
 
-#define KByteToWord(kb) (kb * 1024 / 4)
+#define KByteToWord(kb) ((int)(kb * 1024 / 4))
 #define uxTaskGetFreeStackByte(TaskHandle_x) (uxTaskGetStackHighWaterMark(TaskHandle_x)*4)
